requirements.txt
django
djangorestframework
asgiref
django-cors-headers
djangorestframework-simplejwt
PyJWT
pytz
sqlparse
psycopg
psycopg2-binary
python-dotenv
django-storages
pillow
gunicorn
cloudinary
django-cloudinary-storage
PyMuPDF

